Change default hook values to :options, and circumvent read-only issues
Previously the read-only hook functions were sequence-dependent: it was
important that making a buffer read-only was the last so-long-hook action,
and that reverting that change was the first so-long-revert-hook action.

Ensuring that sequence in customizable hook variables was a problem, but
now we simply use `inhibit-read-only' to sidestep the issue entirely.

Using :type 'hook and and :options list means that there is no longer any
need to use `custom-add-frequent-value' with the defaults.  This also ties
into the aforementioned issue wih ensuring that the read-only functions
happen at particular points in the hook functions sequence, as 'frequent'
values (or :options) are always listed ahead of any custom values.