Use narrowing to optimise `so-long-line-detected-p' performance
Experimentally, `comment-forward' can be slow when invoked at the
start of a very long line, even if there is no comment.  Narrowing
allows us to mitigate this issue.

Similarly, in order to test for a line length > threshold, we need
only test the next (threshold + 1) characters from the start of the
line.  If we do not encounter a newline by then, we know that the
line is too long -- the extent of the excess is irrelevant.