Add `so-long-file-local-mode-function'
Also make function vars automatically buffer-local when set

This makes best sense for `so-long-revert-function' at minimum as, for
any given buffer, it's important that `so-long-revert' calls the revert
function which was configured when `so-long' was triggered.

It also avoids setting the global value inadvertantly, which is useful
given the different ways in which we can trigger `so-long' in combination
with `so-long-file-local-mode-function' potentially modifying these
values.