Change `so-long-change-major-mode' to :before advice for `so-long-mode'
This way we only run the code when we are calling `so-long-mode'.

Previously this was run in `change-major-mode-hook', which meant that
it was redundant during most calls (as most major mode changes are not
changing to `so-long-mode').

Consequently we needn't test `this-command' (which was less rigorous
than knowing that `so-long-mode' was being called somehow), and can
instead use `so-long--calling' to decide how to behave.

This has a flow-on effect in `so-long-mode' itself, so we no longer
need to reset `so-long-original-values' there.

When Emacs bug 35351 is resolved we will use that new feature in Emacs
versions 27+, in place of the advice.