Use (interactive "@") for menu commands
This renders `so-long-menu-click-window' unnecessary, as it is almost
equivalent to (with-selected-window (so-long-menu-click-window) ...)

The minor difference is that the new code leaves the window selected,
whereas the old code did not; but this is of no concern.

The old code should have used (posn-window (event-start last-nonmenu-event)
instead of (car (cadr last-nonmenu-event)), but that was the only such
instance in the library, and this commit removes it entirely.

Co-authored-by: Phil Sainty <psainty@orcon.net.nz>